requirements.txt
PyGithub==2.1.1
github-actions-utils==1.0.1
